vigyani_inventory/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
# gunicorn==21.2.0

# Database
PyMySQL==1.1.0


# Authentication
Authlib==1.3.0

# HTTP client
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
# razorpay
razorpay==1.4.2